sensor_producer/requirements.txt
pika
kafka-python